Added editBug from wangyangjun/node-fogbugz patch-2
Inserted editBug & related from branch

https://github.com/wangyangjun/node-fogbugz/commit/3797cb7993c81825d166517b1766b56dc641ed33

Updated bug in spec (no functional change)

Added MODULE_ERROR for bug not fund

Added bug_not_found and updated use in search.
Changed the editBug to return xml parse error, since that is really what
it is returning, it's not checking for a case (would be nice).

Added Tests for editBug method

Pretty minimal, but we're testing that we'll get an error if there is
not result.
And testing that we'll get the case that we expected to be working on.

Updated Parse to pass in the defferred.

Otherwise it depends on the context which may change.